Add HTTP, HTTPS, FTP Support to Protect (Resolves #149) Resolves #149
This change adds a method to common.py that has the ability to take in
an url and download the file it leads to. In technicality this should be able
to parse any link given to it. However only HTTPS, HTTP, and FTP are being
tested in the test files. If the link is faulty it will throw an error.
If the link is for S3 then it will try downloading the file with the
download _from_s3 method if it can’t with the new method.